Add support for four game modes
On the home screen, a user can select whether each player is a human
or a computer. When the user starts the game, the game will proceed either
as a Human vs. Human, Human vs. Computer, Computer vs. Human, or Computer
vs. Computer game.

This uses ES6 classes and inheritance to create polymorphic functions
for each of the different game modes. A GameFactory is used to determine
which game class to instantiate, and a GameState is used to keep track
of a game's attributes.