Add Handlebars templates
This version uses handlebars to import the menu and board templates from
a separate HTML file. Files are loaded using ajax (a local server is
required to run locally) and compiled using Handlebars.

In order for this to work, I extracted a TicTacToe class with a main
function which is the callback after the partials are successfully
loaded. This ensures that event listeners are only added after the
entire page is loaded.